requirements.txt
alembic==1.9.3
aniso8601==9.0.1
astroid==2.14.2
attrs==22.2.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
Flask==2.2.2
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
flask-serialize==2.1.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.32
mysqlclient==2.1.1
permissive-dict==1.0.4
platformdirs==3.0.0
protobuf==3.20.3
pylint==2.16.2
pyrsistent==0.19.3
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
waitress==2.1.2
Werkzeug==2.2.3
wrapt==1.14.1
